assign node 1

+++ include/window.h
@@ -15,7 +15,7 @@ struct WindowElement {
 
 struct Window {
 	int x, y, width, height; // TODO: int/float?
-	float padding, spacing;
+	int padding, spacing;
 	std::vector<WindowElement> elements;
 	Color bgColor;
 	bool focused;
@@ -26,8 +26,8 @@ struct Window {
 Window* NewWindow();
 void SetWindowColor(Window* win, Color bgColor);
 void SetWindowPosition(Window* win, int x, int y);
-void SetWindowPadding(Window* win, float padding);
-void SetWindowSpacing(Window* win, float spacing);
+void SetWindowPadding(Window* win, int padding);
+void SetWindowSpacing(Window* win, int spacing);
 void AddElementToWindow(Window* win, WindowElement el);
 void DrawWindow(Window* win);
 bool IsWindowFocused(Window* win);
